# Conflicts:
#	src/SFA.DAS.EmployerIncentives.Commands/CreateIncentiveApplication/CreateIncentiveApplicationCommandHandler.cs
#	src/SFA.DAS.EmployerIncentives.Domain.UnitTests/Factories/IncentiveApplicationFactoryTests/WhenCreatedApprenticeship.cs
#	src/SFA.DAS.EmployerIncentives.Domain.UnitTests/IncentiveApplicationTests/WhenAnApprenticeshipIsCreated.cs
#	src/SFA.DAS.EmployerIncentives.Domain/Factories/IIncentiveApplicationFactory.cs
#	src/SFA.DAS.EmployerIncentives.Domain/Factories/IncentiveApplicationFactory.cs
#	src/SFA.DAS.EmployerIncentives.UnitTests.Shared/AutoFixtureCustomizations/IncentiveApplicationCustomization.cs
#	src/tests/SFA.DAS.EmployerIncentives.Commands.UnitTests/CreateIncentiveApplication/Handlers/WhenHandlingCreateIncentiveApplicationCommand.cs

+++ src/SFA.DAS.EmployerIncentives.Abstractions/DTOs/Queries/IncentiveApplicationApprenticeshipDto.cs
@@ -2,7 +2,7 @@
 {
     public class IncentiveApplicationApprenticeshipDto
     {
-        public int ApprenticeshipId { get; set; }
+        public long ApprenticeshipId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public decimal TotalIncentiveAmount { get; set; }
